meta-balena-up-board:layer.conf: Pin kernel at 4.14.76
The BSP has this set to AUTOREV for Up Board. Because we
currently use the Up Board image for all the Up boards,
we get the autorev'ed kernel on the Up Core also and that
newer kernel does not play nice with the wifi chip on the
Up Core. So we pin the Up Board kernel to version 4.14.76
(the same kernel version the BSP has for the Up Core machine)

Changelog-entry: Pin the Up Board kernel to version 4.14.76
Signed-off-by: Florin Sarbu <florin@balena.io>