resin-image: Add grubenv and remove kernel from the boot partition
We can load the kernel from the root partition so let's not deploy
it to the boot partition as we won't load it from there anymore.
We also add the grubenv to the boot partition so we can use the
bootcount mechanism needed for rollbacks.

Changelog-entry: Add grubenv and remove kernel from the boot partition
Signed-off-by: Florin Sarbu <florin@balena.io>