hostapp-update-hooks: Add hook to clean up kernel in boot partition
The kernel in the root partition in the /boot folder is now used to
boot resinOS. After a HUP, there will be a redundant vmlinuz file
left in the boot partition.

This hooks tests for its presence and removes the redundant file.

This commit is taken from Zubair's commit in balena-intel:
https://github.com/balena-os/balena-intel/commit/5063e070

Changelog-entry: Add hook to remove kernel from boot partition after a HUP
Signed-off-by: Florin Sarbu <florin@balena.io>